+++ internal/discovery/git_blame.go
@@ -38,6 +38,8 @@ type GitBlameFinder struct {
 }
 
 func (f GitBlameFinder) Find() (entries []Entry, err error) {
+	slog.Info("Finding all rules to check on current git branch using git blame", slog.String("base", f.baseBranch))
+
 	cr, err := git.CommitRange(f.gitCmd, f.baseBranch)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get the list of commits to scan: %w", err)